Add printf() exactly as it appears in section 9.3
The B implementation of printf works as described with one caveat:
only the 1st argument and all arguments after the 6th will be printed.

This printf implementation relies on pointer arithmetic and assumes
all arguments are placed consecutively on the stack; however, on x86_64,
the first 6 arguments are passed in registers.

In other words,

	printf("I can count! %d, %d, %d, %d, %d, %d, %d, %d, %d*n",
		1, 2, 3, 4, 5, 6, 7, 8, 9);

yields something like:

	I can count! 1, 6295584, 0, 7, 4195265, 6, 7, 8, 9

It's also worth noting that, despite the function definition having ten
arguments (the format string + nine others), printf will actually work
with any number of arguments (more or less than ten, so long as the
number of format specifiers matches).